fix(lint): Enforce React Fast Refresh boundaries
Keep the component-only export rule enabled for TypeScript application files. Separate route definitions, providers, helpers, and variants so existing modules comply without broad lint exceptions.
