+++ src/client/routes/__root.tsx
@@ -1,69 +1,12 @@
-import { useEffect } from "react";
-import { PostHogProvider } from "@posthog/react";
+import { createRootRoute } from "@tanstack/react-router";
 import {
-  Outlet,
-  createRootRoute,
-  type ErrorComponentProps,
-} from "@tanstack/react-router";
-import { getClientConfig } from "@/client/config";
-import { Sentry } from "@/client/observability";
-
-function RootError({ error, reset }: ErrorComponentProps) {
-  useEffect(() => {
-    Sentry.captureException(error);
-  }, [error]);
-
-  return (
-    <main className="flex min-h-screen items-center justify-center bg-background p-6 text-foreground">
-      <div className="max-w-md space-y-4 text-center">
-        <h1 className="text-2xl font-semibold">Something went wrong</h1>
-        <p className="text-sm text-muted-foreground">
-          illiniSpots could not load this view. The error has been reported.
-        </p>
-        <button
-          type="button"
-          onClick={reset}
-          className="rounded-md bg-primary px-4 py-2 text-sm font-medium text-primary-foreground"
-        >
-          Try again
-        </button>
-      </div>
-    </main>
-  );
-}
-
-function RootComponent() {
-  const config = getClientConfig();
-
-  if (!config.posthogProjectToken || !config.posthogHost) {
-    return <Outlet />;
-  }
-
-  return (
-    <PostHogProvider
-      apiKey={config.posthogProjectToken}
-      options={{
-        api_host: config.posthogHost,
-        defaults: "2026-01-30",
-        debug: import.meta.env.DEV,
-      }}
-    >
-      <Outlet />
-    </PostHogProvider>
-  );
-}
+  RootComponent,
+  RootError,
+  RootNotFound,
+} from "@/client/RootRouteComponents";
 
 export const Route = createRootRoute({
   component: RootComponent,
   errorComponent: RootError,
-  notFoundComponent: () => (
-    <main className="flex min-h-screen items-center justify-center bg-background p-6 text-foreground">
-      <div className="space-y-2 text-center">
-        <h1 className="text-2xl font-semibold">Page not found</h1>
-        <a className="text-sm underline" href="/">
-          Return to illiniSpots
-        </a>
-      </div>
-    </main>
-  ),
+  notFoundComponent: RootNotFound,
 });
