+++ src/components/ui/calendar.tsx
@@ -14,7 +14,7 @@ import {
 } from "date-fns";
 import { ChevronLeft, ChevronRight } from "lucide-react";
 import { cn } from "@/lib/utils";
-import { buttonVariants } from "@/components/ui/button";
+import { buttonVariants } from "@/components/ui/buttonVariants";
 
 export interface CalendarProps {
   className?: string;
