+++ src/contexts/theme.ts
@@ -1,0 +1,65 @@
+import { createContext, useContext } from "react";
+
+export type Theme = "light" | "dark" | "system";
+export type ResolvedTheme = "light" | "dark";
+
+export interface ThemeContextType {
+  theme: Theme;
+  resolvedTheme: ResolvedTheme;
+  setTheme: (theme: Theme) => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined,
+);
+
+export const THEME_STORAGE_KEY = "theme";
+const MEDIA_QUERY = "(prefers-color-scheme: dark)";
+
+export function subscribeSystemTheme(callback: () => void) {
+  if (typeof window === "undefined") return () => {};
+  const mq = window.matchMedia(MEDIA_QUERY);
+  mq.addEventListener("change", callback);
+  return () => mq.removeEventListener("change", callback);
+}
+
+export function getSystemThemeSnapshot(): ResolvedTheme {
+  if (typeof window === "undefined") return "light";
+  return window.matchMedia(MEDIA_QUERY).matches ? "dark" : "light";
+}
+
+export function getInitialTheme(): Theme {
+  if (typeof window === "undefined") return "system";
+  try {
+    const stored = localStorage.getItem(THEME_STORAGE_KEY) as Theme | null;
+    if (stored === "light" || stored === "dark" || stored === "system") {
+      return stored;
+    }
+  } catch {
+    // Continue with default fallback if storage is inaccessible
+  }
+  return "system";
+}
+
+export function applyThemeToDocument(resolved: ResolvedTheme) {
+  if (typeof document === "undefined") return;
+
+  const root = document.documentElement;
+  root.classList.toggle("dark", resolved === "dark");
+
+  const metaThemeColor = document.querySelector('meta[name="theme-color"]');
+  if (metaThemeColor) {
+    metaThemeColor.setAttribute(
+      "content",
+      resolved === "dark" ? "#09090b" : "#13294b",
+    );
+  }
+}
+
+export const useTheme = (): ThemeContextType => {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error("useTheme must be used within a ThemeProvider");
+  }
+  return context;
+};
